+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   - fixes [Issue #413](https://github.com/PowerShell/NetworkingDsc/issues/413).
 - Corrected minor style and consistency issues in `NetworkingDsc.Common.tests.ps1`
   and `NetworkingDsc.Common.ps1`.
+- Changed verbose messages in `Test-DscParameterState` to include
+  full type name.
 
 ## 7.3.0.0